refactor: simplify build system by removing unused distribution tasks
Remove distribution-related Gradle tasks that were not used in actual
CI workflows. GitHub Actions handles multi-platform builds directly
using cargo commands, making these Gradle abstractions unnecessary.

Changes:
- Remove NativeLibraryTasks.kt (copyNativeLibs, cleanNativeLibs tasks)
- Replace PlatformConfig.kt with simplified PlatformUtils.kt
- Update NativeBuildTasks.kt to focus on local development only
- Remove cleanNativeLibs dependency from build.gradle.kts
- Clarify separation of concerns in DEVELOPMENT.md:
  * Local: Debug builds for development/testing
  * CI: Release builds handled by GitHub Actions

This simplifies the build system and eliminates duplicate logic between
Gradle and GitHub Actions workflows.
